Set slice from redux-toolkit
