+++ src/index.jsx
@@ -3,11 +3,12 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import { App } from './components/app/app';
 import ErrorBoundary from './components/error-boundary/error-boundary';
-import { Provider } from 'react-redux';import reportWebVitals from './reportWebVitals';
+import { Provider } from 'react-redux';
+import reportWebVitals from './reportWebVitals';
 
 import { configureStore } from '@reduxjs/toolkit'
 import logger from 'redux-logger'
-import { rootReducer } from './services/reducers/index';
+import { rootReducer } from './services/index';
 
 
 const store = configureStore({
