--- src/services/actions/burger.js
@@ -1,4 +1,0 @@
-export const CLEAR = "CLAER";
-export const RANDOM = "RANDOM";
-export const ADD = "ADD";
-export const DELETE = "DELETE";
