+++ src/services/actions/foods.js
@@ -1,23 +1,23 @@
 
-import { loadFoods } from "../../utils/api";
+// import { loadFoods } from "../../utils/api";
+// import { createAction } from '@reduxjs/toolkit'
 
-export const FOODS_LOADING = "FOODS_LOADING";
-export const FOODS_SUCCESS = "FOODS_SUCCESS";
-export const FOODS_ERROR = "FOODS_ERROR";
+// export const FOODS_LOADING = "FOODS_LOADING";
+// export const FOODS_SUCCESS = "FOODS_SUCCESS";
+// export const FOODS_ERROR = "FOODS_ERROR";
 
+// const foodsLoading = createAction(FOODS_LOADING);
+// const foodsError = createAction(FOODS_ERROR);
+// const foodsSuccess = createAction(FOODS_SUCCESS);
 
-export function getFoods () {
-    return function (dispatch) {
-        dispatch({type: FOODS_LOADING});
-        loadFoods()
-            .then(result => dispatch({
-                type: FOODS_SUCCESS,
-                items: result.data
-            })
-            )
-            .catch(e => {
-                console.log("Error:", e);
-                dispatch({ type: FOODS_ERROR})
-            })
- }
-}
+// export function getFoods () {
+//     return function (dispatch) {
+//         dispatch(foodsLoading);
+//         loadFoods()
+//             .then(result => dispatch(foodsSuccess({items: result.data})))
+//             .catch(e => {
+//                 console.log("Error:", e);
+//                 dispatch(foodsError)
+//             })
+//  }
+// }
