+++ src/services/index.js
@@ -1,8 +1,6 @@
 import { combineReducers } from "redux";
-import { burgerReducer } from "./burger";
-import { foodsReducer } from './foods';
+import { foodsReducer } from './slicers/foods';
 
 export const rootReducer = combineReducers({
-    burger: burgerReducer,
     foods: foodsReducer
 });
